+++ Data/shaders/FilmGrain.hlsl
@@ -35,7 +35,7 @@ float gaussian(float z, float u, float o) {
 [numthreads(thread_group_count, thread_group_count, 1)]
 void mainCS(uint3 thread_id : SV_DispatchThreadID)
 {
-    if (thread_id.x >= g_resolution.x || thread_id.y >= g_resolution.y)
+    if (thread_id.x >= uint(g_resolution.x) || thread_id.y >= uint(g_resolution.y))
         return;
     
     const float2 uv = (thread_id.xy + 0.5f) / g_resolution;
